requirements.txt
networkx==2.6.2
osmnx==1.1.1
coverage==5.5.0
shapely==1.8.4
pytest
flake8
networkx_viewer
PyQt5
scikit-learn==1.1.2
ffmpeg-python
opencv-contrib-python
pydot
pandas==1.1.2
numpy==1.23.1
python-slugify==6.1.2

